project/healthbot/requirements.txt
langchain==0.2.16
langchain_openai==0.1.23
langgraph==0.2.19
langchainhub==0.1.21
tavily-python==0.4.0
langchain-community==0.2.16
python-dotenv==1.0.1
